Implements enemy and VHS health system.
Implements a health system for both the enemies and the VHS, including a health bar UI element for the VHS.

Refactors enemy movement into a health-aware basic enemy script, and modifies the enemy spawner to instantiate and initialize enemies with health and target data.

Changes bullet damage to be a positive value to align with adding to health.

+++ Assets/_Project/Scripts/Core/HealthManagement/IDamageable.cs
@@ -2,6 +2,6 @@ namespace _Project.Scripts.Core.HealthManagement
 {
     public interface IDamageable
     {
-        void Damage(int damage);
+        void Damage(float damage);
     }
 }
